django_backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Django
mongoengine
bcrypt
